@@ -7,8 +7,8 @@
     "browserslist": "^4.16.5",
     "glob-parent": "^5.1.2",
     "immer": "9.0.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "4.0.3",
     "set-value": "4.0.1",
     "typescript": "^4.4.3"
